Changed the logo image file name from "Logo.png" to "logo.png" in the login page HTML.
Explanation:
The code diff shows that there is a change in the file name of the logo image in the login page HTML. The file name has been changed from "Logo.png" to "logo.png". This change may have been made to correct the case sensitivity of the file name or to update the logo image with a new version. However, the exact reason for this change is not clear from the code diff itself.